Support compiling for x86_64 Linux

+++ optimize.c
@@ -1116,7 +1116,7 @@ static void GetNamelistData(char* namelist, int* outEntries, char** outEndptr, i
 	//first find the offsets
 	list = namelist + 4;
 	for (i = 0; i < entries; i++) {
-		offsets[i] = 2 + (unsigned int)list - (unsigned int)namelist;
+		offsets[i] = 2 + (unsigned int)(list - namelist);
 		list += *(unsigned short*)list + 2;
 	}
 	memcpy(transforms, offsets, entries * 4);
